backend: Change renderer interface
Use generate_pdf::generate(connection, id) to generate a PDF of the report
with the given id, obtaining data from the database instead of reading
the processed data from a CSV file. If everything worked out ok, generate()
returns the path to the resulting PDF file. This is then served via
rocket::NamedFile.